Moved Proxy within dilia.
