+++ editor/modules/runner/lycan.service.ts
@@ -1,128 +1,34 @@
-import * as io from 'socket.io-client';
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs/Observable';
 import {Subscription} from 'rxjs/Subscription';
 import {Subject} from 'rxjs/Subject';
 
 import {Player} from './player';
+import {LycanCommand, LycanMessage, Direction} from '../../shared';
+import {LycanCommandAuthenticate, ThisIsYou, GameUpdate, LycanEntityUpdate} from '../../shared';
+import {SocketIOService} from '../../services/index';
 
-// Definition of messages
-export interface Point {
-    x: number;
-    y: number;
-}
-
-export const enum Direction {
-    UP,
-    DOWN,
-    LEFT,
-    RIGHT,
-}
-
-export interface LycanEntityUpdate {
-    entity_id: number;
-    position: Point;
-    speed: Point;
-    pv: number;
-}
-
-export interface GameUpdate {
-    kind: 'GameUpdate';
-    entities: LycanEntityUpdate[];
-}
-
-export interface ThisIsYou {
-    kind: 'ThisIsYou';
-    entity: number;
-}
-
-export interface Response {
-    kind: 'Response';
-    code: number;
-}
-
-export interface NewEntity {
-    kind: 'NewEntity';
-    entity: number;
-    position: Point;
-    skin: number;
-    pv: number;
-    nominal_speed: number;
-}
-
-export interface EntityHasQuit {
-    kind: 'EntityHasQuit';
-    entity: number;
-}
-
-export interface Damage {
-    kind: 'Damage';
-    source: number;
-    victim: number;
-    amount: number;
-}
-
-export interface Death {
-    kind: 'Death';
-    entity: number;
-}
-
-export type LycanMessage = GameUpdate
-    | Response
-    | ThisIsYou
-    | NewEntity
-    | EntityHasQuit
-    | Damage
-    | Death;
-
-
-export interface LycanCommandAuthenticate {
-    kind: 'Authenticate';
-    guid: string;
-    token: string;
-}
-
-export interface LycanOrderWalk {
-    kind: 'Walk';
-    entity: number;
-    direction?: Direction;
-}
-
-export interface LycanOrderAttack {
-    kind: 'Attack';
-    entity: number;
-}
-
-export type LycanCommand = LycanCommandAuthenticate
-    | LycanOrderWalk
-    | LycanOrderAttack;
 
 @Injectable()
 export class LycanService {
     private input_stream: Subject<LycanMessage> = new Subject<LycanMessage>();
     private output_stream: Subject<LycanCommand> = new Subject<LycanCommand>();
     private output_sub: Subscription;
-    private socket: SocketIOClient.Socket;
 
-    constructor(private player: Player) {}
+    constructor(
+        private player: Player,
+        private socket: SocketIOService
+    ) {}
 
     connectToLycan() {
-        if (this.socket) {
-            this.socket.disconnect();
-            this.output_sub.unsubscribe();
-        }
-        this.socket = io('ws://localhost:9010', {
-            path: '/lycan',
-        });
-        this.socket.on('message', (message: string) => {
-            let parsed = parse(message);
-            if (parsed !== undefined) {
-                this.input_stream.next(parsed);
+        this.output_sub = this.socket.dualStream<LycanCommand, LycanMessage | undefined>(
+            '/api/lycan',
+            this.output_stream
+        ).subscribe(message => {
+            if (message !== undefined) {
+                this.input_stream.next(message);
             }
         });
-        this.output_sub = this.output_stream
-            .map(serialize)
-            .subscribe(v => this.socket.send(v));
 
         let authenticate: LycanCommandAuthenticate = {
             kind: 'Authenticate',
@@ -169,107 +75,3 @@ export class LycanService {
     }
 }
 
-interface RawLycanMessage {
-    GameUpdate?: GameUpdate;
-    ThisIsYou?: ThisIsYou;
-    Response?: Response;
-    NewEntity?: NewEntity;
-    EntityHasQuit?: EntityHasQuit;
-    Damage?: Damage;
-    Death?: Death;
-}
-
-function parse(message: string): LycanMessage | undefined {
-    let json: RawLycanMessage = JSON.parse(message);
-    if (json.ThisIsYou) {
-        let ret = json.ThisIsYou;
-        ret.kind = 'ThisIsYou';
-        return ret;
-    }
-    if (json.Response) {
-        let ret = json.Response;
-        ret.kind = 'Response';
-        return ret;
-    }
-    if (json.NewEntity) {
-        let ret = json.NewEntity;
-        ret.kind = 'NewEntity';
-        return ret;
-    }
-    if (json.GameUpdate) {
-        let ret = json.GameUpdate;
-        ret.kind = 'GameUpdate';
-        return ret;
-    }
-    if (json.EntityHasQuit) {
-        let ret = json.EntityHasQuit;
-        ret.kind = 'EntityHasQuit';
-        return ret;
-    }
-    if (json.Damage) {
-        let ret = json.Damage;
-        ret.kind = 'Damage';
-        return ret;
-    }
-    if (json.Death) {
-        let ret = json.Death;
-        ret.kind = 'Death';
-        return ret;
-    }
-
-    console.log(`Warning: could not parse ${message}`);
-}
-
-function serialize(command: LycanCommand): string {
-    let res: any;
-    switch (command.kind) {
-        case 'Authenticate': {
-            res = {
-                GameCommand: {
-                    Authenticate: [command.guid, command.token],
-                }
-            };
-            break;
-        }
-
-        case 'Walk': {
-            // WHY did I take North/South/East/West? It is stupid ...
-            // :D
-            let direction: string | null = null;
-            switch (command.direction) {
-                case Direction.UP:
-                    direction = 'North';
-                    break;
-                case Direction.DOWN:
-                    direction = 'South';
-                    break;
-                case Direction.LEFT:
-                    direction = 'West';
-                    break;
-                case Direction.RIGHT:
-                    direction = 'East';
-                    break;
-            }
-            res = {
-                EntityOrder: {
-                    entity: command.entity,
-                    order: {
-                        Walk: direction,
-                    }
-                }
-            };
-            break;
-        }
-
-        case 'Attack': {
-            res = {
-                EntityOrder: {
-                    entity: command.entity,
-                    order: 'Attack',
-                }
-            };
-            break;
-        }
-    }
-    return JSON.stringify(res);
-}
