+++ server/shared.ts
@@ -1,3 +1,4 @@
 export * from '../shared/aariba';
 export * from '../shared/socket';
 export * from '../shared/map';
+export * from '../shared/lycan';
